@porkybrain 2016-08-23
Nice job! I think we can optimize even further by using the "universal fast track approach". still 500 elements in just over a second is amazing!