@jonathanpoelen 2016-08-23
> I think we can optimize even further by using the "universal fast track approach".

Probably. I tried some tricks, but without results. I think tried 1 or 2 other things.
Next step, less than a second with 500 elements? :D