@jonathanpoelen 2016-08-24
I just test again with the "universal fast track" by applying it directly on `detail::apply`:

```cpp
        template <template<typename,typename> class F, typename T, typename U>
        struct apply<F<_1,_2>, list<T, U>> : F<T, U>
        {
        };

        template <template<typename,typename> class F, typename T, typename U>
        struct apply<F<_2,_1>, list<T, U>> : F<U, T>
        {
        };
```

gcc: ~+20%
clang: ~+6%