@porkybrain 2016-08-24
Yes under one second, the race is on ;) for kvasir I need 1000 elements to not crash and not be more than the default max instantiation depth