@ldionne 2016-08-24
Is this sort stable?