@jonathanpoelen 2016-08-24
> Is this sort stable?

No, but it could.

In the meantime.
```cpp
template<class L, class Comp> using stable_sort
  = sort<L, not_<bind<apply, pin<Comp>, _2, _1>>>;
```