@edouarda 2016-08-25
@ldionne do you have any specific use case in mind for a stable sort? I'm curious.