@ldionne 2016-08-25
It's just that Hana's sort is documented as being stable, and I would love to steal Brigand's implementation.