@ldionne 2016-08-25
Well, if anything, it would be wiser to provide a `stable_sort` than to mandate the current `sort` to be stable; this way you don't lock yourself into a corner where you _have_ to make your `sort` function stable (as I did).