@brunocodutra 2016-08-27
`metal::sort` is stable and, as of brunocodutra/metal@72115c21e0e77380529e463de0a72e2698833e2a, [pretty fast too](https://github.com/edouarda/brigand/issues/194#issuecomment-242886273) ;-)